+++ src/Arborescence.Traversal/EnumerableDfs/EnumerableDfs.EnumerateVertices.SingleSource.cs
@@ -1,9 +1,10 @@
 namespace Arborescence.Traversal
 {
+    using System;
     using System.Collections.Generic;
 
     public readonly partial struct EnumerableDfs<
-        TGraph, TVertex, TEdge, TEdgeEnumerator, TExploredSet, TGraphPolicy, TExploredSetPolicy>
+        TGraph, TVertex, TEdge, TEdgeEnumerator, TExploredSet, TExploredSetPolicy>
     {
         // https://11011110.github.io/blog/2013/12/17/stack-based-graph-traversal.html
 
@@ -14,14 +15,20 @@ namespace Arborescence.Traversal
         /// <param name="source">The source.</param>
         /// <param name="exploredSet">The set of explored vertices.</param>
         /// <returns>An enumerator to enumerate the vertices of the the graph.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="graph"/> is <see langword="null"/>.
+        /// </exception>
         public IEnumerator<TVertex> EnumerateVertices(TGraph graph, TVertex source, TExploredSet exploredSet)
         {
+            if (graph == null)
+                throw new ArgumentNullException(nameof(graph));
+
             var stack = new Internal.Stack<TEdgeEnumerator>();
             try
             {
                 ExploredSetPolicy.Add(exploredSet, source);
                 yield return source;
-                stack.Add(GraphPolicy.EnumerateOutEdges(graph, source));
+                stack.Add(graph.EnumerateOutEdges(source));
 
                 while (stack.TryTake(out TEdgeEnumerator outEdges))
                 {
@@ -31,7 +38,7 @@ namespace Arborescence.Traversal
                     stack.Add(outEdges);
 
                     TEdge e = outEdges.Current;
-                    if (!GraphPolicy.TryGetHead(graph, e, out TVertex v))
+                    if (!graph.TryGetHead(e, out TVertex v))
                         continue;
 
                     if (ExploredSetPolicy.Contains(exploredSet, v))
@@ -39,7 +46,7 @@ namespace Arborescence.Traversal
 
                     ExploredSetPolicy.Add(exploredSet, v);
                     yield return v;
-                    stack.Add(GraphPolicy.EnumerateOutEdges(graph, v));
+                    stack.Add(graph.EnumerateOutEdges(v));
                 }
             }
             finally