+++ src/Arborescence.Traversal/InstantBfs/InstantBfs_5.cs
@@ -3,8 +3,8 @@ namespace Arborescence.Traversal
     using System.Collections.Generic;
 
     /// <summary>
-    /// Provides the <see cref="Create{TGraphPolicy,TColorMapPolicy}"/> factory method
-    /// for <see cref="InstantBfs{TGraph,TVertex,TEdge,TEdgeEnumerator,TColorMap,TGraphPolicy,TColorMapPolicy}"/>.
+    /// Provides the <see cref="Create{TColorMapPolicy}"/> factory method
+    /// for <see cref="InstantBfs{TGraph,TVertex,TEdge,TEdgeEnumerator,TColorMap,TColorMapPolicy}"/>.
     /// </summary>
     /// <typeparam name="TGraph">The type of the graph.</typeparam>
     /// <typeparam name="TVertex">The type of the vertex.</typeparam>
@@ -12,26 +12,23 @@ namespace Arborescence.Traversal
     /// <typeparam name="TEdgeEnumerator">The type of the edge enumerator.</typeparam>
     /// <typeparam name="TColorMap">The type of the vertex color map.</typeparam>
     public static class InstantBfs<TGraph, TVertex, TEdge, TEdgeEnumerator, TColorMap>
+        where TGraph : IOutEdgesConcept<TVertex, TEdgeEnumerator>, IHeadConcept<TVertex, TEdge>
         where TEdgeEnumerator : IEnumerator<TEdge>
     {
         /// <summary>
-        /// Creates a new <see cref="InstantBfs{TGraph,TVertex,TEdge,TEdgeEnumerator,TColorMap,TGraphPolicy,TColorMapPolicy}"/>
+        /// Creates a new <see cref="InstantBfs{TGraph,TVertex,TEdge,TEdgeEnumerator,TColorMap,TColorMapPolicy}"/>
         /// algorithm from the given policies.
         /// </summary>
-        /// <param name="graphPolicy">The graph policy.</param>
         /// <param name="colorMapPolicy">
         /// The <see cref="IMapPolicy{TMap,TKey,TValue}"/> implementation to use when marking explored vertices while traversing.
         /// </param>
-        /// <typeparam name="TGraphPolicy">The type of the graph policy.</typeparam>
         /// <typeparam name="TColorMapPolicy">The type of the vertex color map policy.</typeparam>
         /// <returns>An algorithm instance with specified policies.</returns>
-        public static InstantBfs<TGraph, TVertex, TEdge, TEdgeEnumerator, TColorMap, TGraphPolicy, TColorMapPolicy>
-            Create<TGraphPolicy, TColorMapPolicy>(TGraphPolicy graphPolicy, TColorMapPolicy colorMapPolicy)
-            where TGraphPolicy : IOutEdgesPolicy<TGraph, TVertex, TEdgeEnumerator>, IHeadPolicy<TGraph, TVertex, TEdge>
+        public static InstantBfs<TGraph, TVertex, TEdge, TEdgeEnumerator, TColorMap, TColorMapPolicy>
+            Create<TColorMapPolicy>(TColorMapPolicy colorMapPolicy)
             where TColorMapPolicy : IMapPolicy<TColorMap, TVertex, Color>
         {
-            return new InstantBfs<TGraph, TVertex, TEdge, TEdgeEnumerator, TColorMap, TGraphPolicy, TColorMapPolicy>(
-                graphPolicy, colorMapPolicy);
+            return new InstantBfs<TGraph, TVertex, TEdge, TEdgeEnumerator, TColorMap, TColorMapPolicy>(colorMapPolicy);
         }
     }
 }