+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.6.0] - 2020-11-10
+### Added
+- Abstractions: `IReadOnlySetPolicy<>` and `IReadOnlyMapPolicy<>` interfaces.
+
+### Removed
+- Abstractions: graph policy interfaces.
+- Models: graph policy implementations.
+
 ## [0.5.0] - 2020-11-03
 ### Added
 - Models: Throwing `Add()` method to builders in addition to non-throwing `TryAdd()`.
@@ -86,7 +94,8 @@
 ### Added
 - Abstractions: The interface for graphs to be examined in a data-structure agnostic fashion.
 
-[Unreleased]: https://github.com/qbit86/arborescence/compare/traversal-0.5.0...HEAD
+[Unreleased]: https://github.com/qbit86/arborescence/compare/abstractions-0.6.0...HEAD
+[0.6.0]: https://github.com/qbit86/arborescence/compare/traversal-0.5.0...abstractions-0.6.0
 [0.5.0]: https://github.com/qbit86/arborescence/compare/models-0.4.2...traversal-0.5.0
 [0.4.2]: https://github.com/qbit86/arborescence/compare/models-0.4.1...models-0.4.2
 [0.4.1]: https://github.com/qbit86/arborescence/compare/models-0.4.0...models-0.4.1